US10: Add bonsai state manager test

+++ app/src/main/java/bonsai/habit/database/entity/Bonsai.kt
@@ -16,7 +16,7 @@ data class Bonsai(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     @ColumnInfo(name = "targetTime")
-    val targetTime: Int,
+    val targetTime: Long,
     @ColumnInfo(name = "age")
     val age: Int,
     @ColumnInfo(name = "healthState")
